Documented and tested wmake macro substitution $(macro:from=to). Fixed foo = bar, fubar=$(foo), $(fubar:r=re) $(<:from=to) fails as OW - unlike MS - rejects $(<) as $< synonym.
[git-p4: depot-paths = "//depot/openwatcom/": change = 28155]

+++ bld/wmake/regress/macros/test.bat
@@ -144,6 +144,38 @@ if errorlevel 1 goto err6
 
 :test7
 
+echo # ---------------------------
+echo #   Macro Test 7
+echo # ---------------------------
+
+wtouch hello.obj hello2.obj
+wtouch hello.c  hello2.c
+%1 -h -ms -m -f MACRO07 -a cc=bwcl386 > tmp.out 2>&1
+diff -b MACRO07.CMP tmp.out
+if errorlevel 1 goto err7
+    @echo # MACRO07 successful
+    goto test8
+:err7
+    @echo ## MACRO ## >> %2
+    @echo Error: MACRO #7 unsuccessful!!! | tee -a %2
+
+:test8
+echo # ---------------------------
+echo #   Macro Test 8
+echo # ---------------------------
+
+%1 -ms -h -f MACRO08 > tmp.out 2>&1
+diff -b MACRO08.CMP tmp.out
+if errorlevel 1 goto err8
+    @echo # MACRO08 successful
+    goto test9
+:err8
+    @echo ## MACRO ## >> %2
+    @echo Error: MACRO #8 unsuccessful!!! | tee -a %2
+
+
+:test9
+
 rm -f hello.obj hello2.obj hello.boo hello.tmp hello.c hello2.c 
 goto done
 :usage
